Make frontend generic
